Fix an intermittent issue in scale down.
Problem: When two scale down operations come together, the first is cancelled, and the second one is suspended waiting for the cancelation of the first one forever.

Cause:  Fiber's resume method only handles NOT_COMPLETED state, but a cancelled fiber is in CANCALLED state.

Solution: Add code to handle CANCELLED state in resume.

Signed-off-by: Dongbo Xiao <dongbo.xiao@oracle.com>