do not fail the whole test if the curl command returns a non-zero status (#1595)
* do not fail the whole test if the curl command returns a non-zero status

we expect if to fail, and want to retry it until it is successful.  it
can fail because the web app has not had time to start up yet.  the
current code throws a runtime exception on a non-zero return code, which
causes this test to fail too fast and incorrectly.

* use vanaja's simpler solution