Attempt to clean up use of bad exec()
This commit attempts to resolve a number of test failures that occur due
to the bad use of the TestUtils.exec() method.  This method throws a
RuntimeException if the return code of the executed command is non-zero.
This method is being used in many places (over 40).  Many of those do
not appear to need or want that behavior, and should instead be using
the ExecCommand.exec() method which executes a command and then returns
the result, without throwing an exception if the command fails.

In this commit, I have done a visual inspection of each use of
TestUtils.exec() and changed them to ExecCommand.exec() where it
appeared to make sense.  This includes every case where the call was
using the return value from the method, and those where it appears clear
from the subsequent logic that a failed command was not expected to
terminate the process.

There were some calls that I left because I could not be sure of the
intent, and these will require more study.

In order to make the code clearer, and make people think twice about
using TestUtils.exec(), which has caused so many failures, I renamed
that method to TestUtils.execOrAbortProcess().  Hopefully that will make
it clearer that that method has a dangerous and unexpected side effect.