Fixed the build application utility
The fix involves
Removing the persistent volumes for mounting application source directory
to the pod , since PV_ROOT is shared only between the Kubernetes nodes and
Jenkins client doesn't have access to the PV_ROOT

Instead of a Kubernetes job the utility will create a WebLogic pod and copies
the source application folder to the pod and builds the application.

Once built the archives directory is copied back to the local file system
for use by the test code.