watcher: stop parsing resource version
As noted in the documentation for the utility methods, parsing the
resource version is explicitly disallowed by the Kubernetes API. The
Kubernetes API server ensures that watch event streams deliver events
once and only once, and always in a total order. This is the critical
feature of the system that allows for users to resume a watch from the
latest resource version they have seen without requiring them to do any
parsing or comparison. This change removes the parsing logic from the
codebase and simplifies the reousrce version tracking routine.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>